refactor: move interface-agnostic substrate into gateway_core
