+++ packages/gateway_core/src/gateway_core/models/pagination.py
