+++ packages/gateway_core/src/gateway_core/versioning.py
