+++ packages/gateway_oss/src/gateway_oss/models/__init__.py
@@ -1,0 +1,1 @@
+from __future__ import annotations
